Now we can use all valid number. We have new ability for randomFillArray: we can randomize random numbers.

+++ Course-work/Course-work/SortMethods.h
@@ -3,17 +3,17 @@
 
 using namespace std;
 
-void selectionSort(int *, int, int);
-void randomFillingArray(int *, int, int, int);
-void showArray(int *, int);
-void shellSort(int *, int, int);
-void down(int *, int, int, int);
-void heapSort(int *, int, int);
+void selectionSort(double *, int, int);
+void randomFillingArray(double *, int, double, double);
+void showArray(double *, int);
+void shellSort(double *, int, int);
+void down(double *, int, int, int);
+void heapSort(double *, int, int);
 istream& openFileINPUT(ifstream &, string);
 istream& readSize(ifstream &, string &, int &);
 ostream& openFileOUTPUT(ofstream &, string);
-void fileFillingArray(string, int *, int);
-void chooseMethod(int *Array, int size);
-void manuallyFillingArray(int *Array, int size);
-int *chooseAction(int &);
-void writeFile(int *Array, int size);
+void fileFillingArray(string, double *, int);
+void chooseMethod(double *Array, int size);
+void manuallyFillingArray(double *Array, int size);
+double *chooseAction(int &);
+void writeFile(double *Array, int size);
